refactor(Web Search): 重命名 provider 工具模块
