+++ services/wisepen-chat-service/src/chat/application/tools/search_tools/web_search/__init__.py
@@ -1,4 +1,4 @@
-from .tools.facades import (
+from .tools.provider_tools import (
     AnySearchSearchTool,
     BaiduQianfanSearchTool,
     ExaSearchTool,
