+++ services/wisepen-chat-service/src/chat/application/tools/search_tools/web_search/tools/__init__.py
@@ -1,5 +1,5 @@
 from .base import BaseWebSearchTool
-from .facades import (
+from .provider_tools import (
     AnySearchSearchTool,
     BaiduQianfanSearchTool,
     ExaSearchTool,
